pyproject.toml
[tool.poetry]
name = "automated-backup-system"
version = "0.0.1"
description = "This is Project to practice automation"
authors = ["Brent-the-carpenter <beh337@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.8.0"
ruff = "^0.6.8"
pyyaml = "^6.0.2"
boto3 = "^1.35.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
